Only download media url with curl if audioservice cannot handle it
Curl downloads of the media url have been failing and shouldn't be
necessary in most circumstances. This passes the media url directly
to the audioservice for playback unless it uses the https protocol
and there are no audioservices that cannot handle this. Eg the
simple audioservice currently cannot.

Once https support is added to that service, no explicit change
would be required here.